+++ Rdmp.UI/Collections/CatalogueCollectionUI.cs
@@ -203,7 +203,6 @@ public partial class CatalogueCollectionUI : RDMPCollectionUI
             //reset to flat view
             tlvCatalogues.RemoveObject(Activator.CoreChildProvider.CatalogueRootFolder);
             tlvCatalogues.AddObjects(Activator.CoreChildProvider.AllCatalogues);
-            tlvCatalogues.RefreshObjects(Activator.CoreChildProvider.AllCatalogues);
             _renderingFlatFiew = true;
         }
         else if(!UserSettings.ShowFlatLists && _renderingFlatFiew)